pcm: Add fallback plugin
PCM fallback plugin works as a proxy to a list of slave PCMs.  It
tries to open each pre-given slave PCM until the stream can be
properly opened or it gets -EGAIN error.  That is, when a stream is
non-existent, it's skipped and the next slave PCM is tried.

This plugin is useful for the cases like HD-audio where only HDMI or
SPDIF might be present, so that it may fall back to the only existing
device automatically.

Like asym plugin, fallback plugin also takes effect only at opening a
PCM stream.  Once when the slave PCM is properly opened, the whole
control is moved to the slave PCM.

Signed-off-by: Takashi Iwai <tiwai@suse.de>